requirements.txt
absl-py==1.2.0
einops==0.4.1
h5py==3.7.0
keopscore==2.1
opt-einsum==3.3.0
pandas==1.4.2 
pytorch-wavelet 
PyWavelets==1.4.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
statsmodels==0.13.2
sympy==1.11.1
peft==0.9.0
transformers==4.30.1
patool==1.15.0
matplotlib==3.8.3
ptflops
